+++ Command/Claymore.cs
@@ -10,7 +10,8 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using GolemUI.Extensions;
+using GolemUI.Utils;
+
 namespace GolemUI.Command
 {
